drawStatus-ok
