+++ src/pages/home/index.tsx
@@ -1,18 +1,13 @@
-import { SideBar } from "../layout/sideBar";
 import { StudentsList } from "./studentsList";
+import { DrawTitle } from "./DrawTitle";
 
-function Home()
-{
+function Home() {
 	return (
-		<div className="flex">
-			{/* <h1>welcome to the Home Page</h1>
-			<p>This is the home page of our application.</p>
-			<Link to="/login" className="text-blue-500 hover:underline">
-				Go to Login
-			</Link> */}
-			{StudentsList()}
+		<div className="flex flex-col">
+			<DrawTitle title={"Lista de estudantes"} />
+			<StudentsList />
 		</div>
 	);
-};
+}
 
 export default Home;
